pipeflow: Lightweight, General-Purpose Data Analysis Pipelines

A lightweight yet powerful framework for building robust data analysis pipelines. With 'pipeflow', you initialize a pipeline with your dataset and construct workflows step by step by adding R functions. You can modify, remove, or insert steps and parameters at any stage, while 'pipeflow' ensures the pipeline's integrity. Overall, this package offers a beginner-friendly framework that simplifies and streamlines the development of data analysis pipelines by making them modular, intuitive, and adaptable.

See also

Author

Maintainer: Roman Pahl roman.pahl@gmail.com

Authors:

Roman Pahl roman.pahl@gmail.com